+++ src/lib/code-sample-preview.ts
@@ -19,10 +19,12 @@ import {
   css,
   property,
   query,
+  PropertyValues,
 } from 'lit-element';
 import {ifDefined} from 'lit-html/directives/if-defined.js';
 import '@material/mwc-icon-button';
 import '@material/mwc-textfield';
+import {CodeSampleProjectElement} from './code-sample-project.js';
 
 /**
  * An HTML preview component consisting of a wrapper around an iframe, and
@@ -31,8 +33,8 @@ import '@material/mwc-textfield';
  *
  * @fires reload - Fired when the user clicks the reload button
  */
-@customElement('code-sample-editor-preview')
-export class CodeSampleEditorPreviewElement extends LitElement {
+@customElement('code-sample-preview')
+export class CodeSamplePreviewElement extends LitElement {
   static styles = css`
     * {
       box-sizing: border-box;
@@ -95,6 +97,22 @@ export class CodeSampleEditorPreviewElement extends LitElement {
   @query('iframe')
   private _iframe!: HTMLIFrameElement;
 
+  /**
+   * The project that this preview is associated with. Either the
+   * `<code-sample-project>` node itself, or its `id` in the host scope.
+   */
+  @property()
+  project: CodeSampleProjectElement | string | undefined = undefined;
+
+  private _project: CodeSampleProjectElement | undefined = undefined;
+
+  async update(changedProperties: PropertyValues) {
+    if (changedProperties.has('project')) {
+      this._findProjectAndRegister();
+    }
+    super.update(changedProperties);
+  }
+
   render() {
     return html`
       <div id="toolbar">
@@ -108,12 +126,42 @@ export class CodeSampleEditorPreviewElement extends LitElement {
     `;
   }
 
+  private _findProjectAndRegister() {
+    const prevProject = this._project;
+    if (this.project instanceof HTMLElement) {
+      this._project = this.project;
+    } else if (typeof this.project === 'string') {
+      this._project =
+        (((this.getRootNode() as unknown) as
+          | Document
+          | ShadowRoot).getElementById(
+          this.project
+        ) as CodeSampleProjectElement | null) || undefined;
+    } else {
+      this._project = undefined;
+    }
+    if (prevProject !== this._project) {
+      if (prevProject) {
+        prevProject._unregisterPreview(this);
+      }
+      if (this._project) {
+        this._project._registerPreview(this);
+      }
+    }
+  }
+
   reload() {
     const iframe = this._iframe;
     iframe.contentWindow?.location.reload();
   }
 
   private _onReloadClick() {
-    this.dispatchEvent(new Event('reload'));
+    this._project?.save();
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'code-sample-preview': CodeSamplePreviewElement;
   }
 }